Add note about OPENDJ-3377 as OpenIG Workaround
We are missing OpenIG Toolkit versions after 4.0.0, and OpenDJ 1.5.x was built against OpenIG Toolkit 4.5.2. As a workaround, we're able to take advantage of the fact that the piece of OpenIG that is needed actually moved into FR commons 21.0.0-alpha-6 and higher. FR switched to using that version of the module in OPENDJ-3377, for the 4.x versions of OpenDJ, so we're just doing the same thing here.