Correct ZIP Packages to Use Raw Product Name
the default behavior for DS 3.x was to use the human-friendly product name to name JAR and ZIP files. FR corrected this in 3.5.1 and 4.x, but we didn't have this in 3.x. I've corrected the issue by introducing a lower-cased version of the name for these files.

NOTE: `product.name` is used internally in the DS build as the name of the product that DS matches upon connection. if the property and all POM references to it are renamed, then DS won't build.

Closes #12.