[#29] Fix Failing Test due to Unwanted Port Reuse
It looks like this test was failing because the kernel is reusing the same port for two different sockets within the same application (despite us not asking for this at all). In JDK 9+, we can control this via `StandardSocketOptions.SO_REUSEPORT`, but that option is not available in JDK 8. As a workaround, setting address reuse on the placeholder socket seems to prevent the subsequent socket from binding to the same port.